+++ src/content/items/Equipment.ts
@@ -1,0 +1,4 @@
+export enum Equipment {
+
+    SoulGem             = 0//FourCC(''),
+}