+++ versions/Changelog.md
@@ -7,3 +7,20 @@
 
 ### Class Changes
 - Add Hell Forge to Artisan
+
+
+## Version 0.2
+
+### Main Changes:
+- Added 4 Felsmithing recipes
+- Colorized Material formatted string
+- Added Maintain Temperature to Hell Forge
+- Added Fel Basin to prospector
+
+## Version 0.3
+
+### Main Changes:
+- TODO: Ore gen
+- Improved player commands -die by looking at "player hero" instead of selection
+- Improved Hand ability to also check "player hero"
+